refactor: extract addNotifications factory function
